GIT_SYNC_BRANCH accepts either a branch or a tag
Currently git-sync uses GIT_SYNC_REV to check-out a tag or an hash, if
one-time is set is not necessary to lookup the hash of the tag because
is enough to use GIT_SYNC_BRANCH to clone the repository.
With this change is possible to clone a repository by a tag or a branch using
GIT_SYNC_BRANCH in case one-time is set to true.