Clone the repo to a subdirectory under --root
The git repo is currently cloned into the --root directory. If any git
command fails, an error file will be generated under --root. This is
problematic because it removes all contents under --root if `git clone`
fails, including the error file. The presence of the error file
indicates a prior failure. If `git clone` is very slow and fails again,
the fact that the error file is deleted incorrectly suggests the
previous clone was successful.

We can't simply skip removing the error file because `git clone` requires
the destination directory to be empty. This commit creates a
subdirectory under --root with the repo name, and clones the repo to the
subdirectory. The error file is peer to the subdirectory. With this
approach, if `git clone` fails, it just deletes the subdirectory, with the
error file is retained. The symlink still points to the worktree, with
an extra subdirectory in the path.

An alternative is to make the error file an absolute path, instead of
relative to --root. However, it is possible that the absolute path is
under --root. Then it requires additional validation or error handling.
Therefore, this approach is not chosen.